TestClientandServer
Can you add the server code to each of the servers and keep it running in the background. Then add the client code to the API/Host server and run it (make sure it is after all 3 server codes are running) to test if they communicate properly. if they do, it should display live data about all of the metrics. I also already added the IP's based on what we spoke about yesterday, so the code will be ready.

Co-Authored-By: AARocher <89544076+AARocher@users.noreply.github.com>